requirements.txt
absl-py==2.1.0
astunparse==1.6.3
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dm-tree==0.1.8
docstring_parser==0.16
etils==1.10.0
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.67.0
h5py==3.12.1
idna==3.10
immutabledict==4.2.0
importlib_resources==6.4.5
joblib==1.4.2
keras==3.6.0
keras-tuner==1.4.7
kiwisolver==1.4.7
kt-legacy==1.0.5
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
mypy-extensions==1.0.0
namex==0.0.8
numba==0.60.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
promise==2.3
protobuf==4.25.5
psutil==6.1.0
pyarrow==17.0.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
shap==0.46.0
simple-parsing==0.1.6
six==1.16.0
slicer==0.0.8
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-datasets==4.9.6
tensorflow-intel==2.16.1
tensorflow-metadata==1.16.1
termcolor==2.5.0
tfds-nightly==4.9.6.dev202410250044
threadpoolctl==3.5.0
toml==0.10.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
zipp==3.20.2
